Move invoice creation navigation out of hook
