+++ client/packages/dashboard/src/widgets/ReplenishmentWidget.tsx
@@ -7,6 +7,7 @@ import {
   RANGE_SPLIT_CHAR,
   RouteBuilder,
   StatsPanel,
+  useNavigate,
   useNotification,
   useToggle,
   Widget,
@@ -29,10 +30,11 @@ import { InternalSupplierSearchModal } from '@openmsupply-client/system';
 import { AppRoute } from '@openmsupply-client/config';
 
 export const ReplenishmentWidget: React.FC<PropsWithChildrenOnly> = () => {
+  const t = useTranslation('dashboard');
   const modalControl = useToggle(false);
   const { error: errorNotification } = useNotification();
-  const t = useTranslation('dashboard');
   const formatNumber = useFormatNumber();
+  const navigate = useNavigate();
   const { data, isLoading, isError, error } = useDashboard.statistics.inbound();
   const {
     data: requisitionCount,
@@ -90,7 +92,15 @@ export const ReplenishmentWidget: React.FC<PropsWithChildrenOnly> = () => {
                 otherPartyId,
               },
               { onError }
-            );
+            ).then(invoiceNumber => {
+              navigate(
+                RouteBuilder.create(AppRoute.Replenishment)
+                  .addPart(AppRoute.InboundShipment)
+                  .addPart(String(invoiceNumber))
+                  .build(),
+                { replace: true }
+              );
+            });
           }}
         />
       ) : null}
