+++ client/packages/invoices/src/InboundShipment/api/hooks/document/useInsertInbound.ts
@@ -1,23 +1,11 @@
-import {
-  useNavigate,
-  useQueryClient,
-  useMutation,
-  RouteBuilder,
-} from '@openmsupply-client/common';
-import { AppRoute } from '@openmsupply-client/config';
+import { useQueryClient, useMutation } from '@openmsupply-client/common';
 import { useInboundApi } from '../utils/useInboundApi';
 
 export const useInsertInbound = () => {
   const queryClient = useQueryClient();
-  const navigate = useNavigate();
   const api = useInboundApi();
   return useMutation(api.insert, {
-    onSuccess: invoiceNumber => {
-      const route = RouteBuilder.create(AppRoute.Replenishment)
-        .addPart(AppRoute.InboundShipment)
-        .addPart(String(invoiceNumber))
-        .build();
-      navigate(route, { replace: true });
+    onSuccess: () => {
       return queryClient.invalidateQueries(api.keys.base());
     },
   });
