+++ client/packages/invoices/src/OutboundShipment/api/hooks/document/useOutboundInsert.ts
@@ -1,23 +1,11 @@
-import {
-  RouteBuilder,
-  useNavigate,
-  useMutation,
-  useQueryClient,
-} from '@openmsupply-client/common';
-import { AppRoute } from '@openmsupply-client/config';
+import { useMutation, useQueryClient } from '@openmsupply-client/common';
 import { useOutboundApi } from './../utils/useOutboundApi';
 
 export const useOutboundInsert = () => {
   const queryClient = useQueryClient();
-  const navigate = useNavigate();
   const api = useOutboundApi();
   return useMutation(api.insert.outbound, {
-    onSuccess: invoiceNumber => {
-      const route = RouteBuilder.create(AppRoute.Distribution)
-        .addPart(AppRoute.OutboundShipment)
-        .addPart(String(invoiceNumber))
-        .build();
-      navigate(route, { replace: true });
+    onSuccess: () => {
       queryClient.invalidateQueries(api.keys.base());
     },
   });
