+++ client/packages/invoices/src/Prescriptions/api/hooks/document/usePrescriptionInsert.ts
@@ -1,23 +1,11 @@
-import {
-  RouteBuilder,
-  useNavigate,
-  useMutation,
-  useQueryClient,
-} from '@openmsupply-client/common';
-import { AppRoute } from '@openmsupply-client/config';
+import { useMutation, useQueryClient } from '@openmsupply-client/common';
 import { usePrescriptionApi } from '../../utils/usePrescriptionApi';
 
 export const usePrescriptionInsert = () => {
   const queryClient = useQueryClient();
-  const navigate = useNavigate();
   const api = usePrescriptionApi();
   return useMutation(api.insert, {
-    onSuccess: invoiceNumber => {
-      const route = RouteBuilder.create(AppRoute.Dispensary)
-        .addPart(AppRoute.Prescription)
-        .addPart(String(invoiceNumber))
-        .build();
-      navigate(route, { replace: true });
+    onSuccess: () => {
       queryClient.invalidateQueries(api.keys.base());
     },
   });
